+++ python/libcasm/xtal/_methods.py
@@ -1,7 +1,7 @@
 import functools
 import math
 from collections import namedtuple
-from typing import Any, Callable, Union
+from typing import Any, Callable, Optional, Union
 
 import numpy as np
 
@@ -211,7 +211,7 @@ StructureAtomInfo = namedtuple(
     "StructureAtomInfo",
     ["atom_type", "atom_coordinate_frac", "atom_coordinate_cart", "atom_properties"],
 )
-""" A namedtuple, used to hold atom info when sorting atoms in a 
+""" A namedtuple, used to hold atom info when sorting, filtering, etc. atoms in a 
 :class:`~_xtal.Structure`.
 
 .. rubric:: Constructor
@@ -228,8 +228,102 @@ atom_coordinate_cart: numpy.ndarray[numpy.float64[3]]
     :func:`~_xtal.Structure.atom_type.atom_coordinate_cart`.
 atom_properties: dict[str, numpy.ndarray[numpy.float64[m]]]
     The continuous properties associated with the atoms, if present, from 
-    :func:`~_xtal.Structure.atom_type.atom_coordinate_frac`.
+    :func:`~_xtal.Structure.atom_type.atom_coordinate_frac`. All atoms
+    must have the same properties with values of the same dimension.
 """
+
+
+def make_structure_atom_info(
+    structure: _xtal.Structure,
+) -> list[StructureAtomInfo]:
+    """Create a list of StructureAtomInfo from a Structure
+
+    Parameters
+    ----------
+    structure: _xtal.Structure
+        The structure to be sorted, filtered, etc. by atom info.
+
+    Returns
+    -------
+    structure_atom_info: list[StructureAtomInfo]
+        A list of StructureAtomInfo.
+
+    """
+
+    atom_type = structure.atom_type()
+    atom_coordinate_frac = structure.atom_coordinate_frac()
+    atom_coordinate_cart = structure.atom_coordinate_cart()
+    atom_properties = structure.atom_properties()
+
+    atoms = []
+    import copy
+
+    for i in range(len(atom_type)):
+        atoms.append(
+            StructureAtomInfo(
+                copy.copy(atom_type[i]),
+                atom_coordinate_frac[:, i].copy(),
+                atom_coordinate_cart[:, i].copy(),
+                {key: atom_properties[key][:, i].copy() for key in atom_properties},
+            )
+        )
+
+    return atoms
+
+
+def make_structure_from_atom_info(
+    lattice: _xtal.Lattice,
+    atoms: list[StructureAtomInfo],
+    global_properties: dict[str, np.ndarray[np.float64]] = {},
+) -> _xtal.Structure:
+    """Create a Structure from a list of StructureAtomInfo
+
+    Parameters
+    ----------
+    lattice: _xtal.Lattice]
+        The lattice for the resulting structure.
+    atoms: list[StructureAtomInfo]
+        A list of StructureAtomInfo. The Cartesian coordinates are used when setting
+        atom coordinates in the resulting structure.
+    global_properties: dict[str, numpy.ndarray[numpy.float64[m, n]]] = {}
+        Continuous properties associated with entire crystal, if present. Keys must be
+        the name of a CASM-supported property type. Values are (m, 1) arrays with
+        dimensions matching the standard dimension of the property type.
+
+    Returns
+    -------
+    structure: _xtal.Structure
+        The resulting structure
+    """
+
+    n_atoms = len(atoms)
+
+    atom_type = [atom.atom_type for atom in atoms]
+    atom_coordinate_cart = np.zeros((3, n_atoms))
+    atom_properties = {}
+
+    for i, atom in enumerate(atoms):
+        if i == 0:
+            for key, value in atom.atom_properties.items():
+                dim = value.shape[0]
+                atom_properties[key] = np.zeros((dim, n_atoms))
+
+        atom_coordinate_cart[:, i] = atom.atom_coordinate_cart
+        for key, value in atom.atom_properties.items():
+            atom_properties[key][:, i] = atom.atom_properties[key]
+
+    atom_coordinate_frac = _xtal.cartesian_to_fractional(
+        lattice=lattice,
+        coordinate_cart=atom_coordinate_cart,
+    )
+
+    return _xtal.Structure(
+        lattice=lattice,
+        atom_type=atom_type,
+        atom_coordinate_frac=atom_coordinate_frac,
+        atom_properties=atom_properties,
+        global_properties=global_properties,
+    )
 
 
 def sort_structure_by_atom_info(
@@ -265,41 +359,15 @@ def sort_structure_by_atom_info(
         raise ValueError(
             "Error: only atomic structures may be sorted using sort_by_atom_info"
         )
-    atom_type = structure.atom_type()
-    atom_coordinate_frac = structure.atom_coordinate_frac()
-    atom_coordinate_cart = structure.atom_coordinate_cart()
-    atom_properties = structure.atom_properties()
 
-    atoms = []
-    import copy
-
-    for i in range(len(atom_type)):
-        atoms.append(
-            StructureAtomInfo(
-                copy.copy(atom_type[i]),
-                atom_coordinate_frac[:, i].copy(),
-                atom_coordinate_cart[:, i].copy(),
-                {key: atom_properties[key][:, i].copy() for key in atom_properties},
-            )
-        )
-
+    atoms = make_structure_atom_info(structure)
     atoms.sort(key=key, reverse=reverse)
 
-    for i, atom in enumerate(atoms):
-        atom_type[i] = atom[0]
-        atom_coordinate_frac[:, i] = atom[1]
-        for key in atom_properties:
-            atom_properties[key][:, i] = atom[2][key]
-
-    sorted_struc = _xtal.Structure(
+    return make_structure_from_atom_info(
         lattice=structure.lattice(),
-        atom_type=atom_type,
-        atom_coordinate_frac=atom_coordinate_frac,
-        atom_properties=atom_properties,
+        atoms=atoms,
         global_properties=structure.global_properties(),
     )
-
-    return sorted_struc
 
 
 def sort_structure_by_atom_type(
@@ -473,4 +541,79 @@ def substitute_structure_species(
         mol_type=[substitutions.get(x, x) for x in structure.mol_type()],
         mol_properties=structure.mol_properties(),
         global_properties=structure.global_properties(),
+    )
+
+
+def filter_structure_by_atom_info(
+    structure: _xtal.Structure,
+    filter: Callable[[StructureAtomInfo], Any],
+) -> _xtal.Structure:
+    """Return a copy of a structure with atoms passing a filter function
+
+    .. rubric:: Example usage
+
+    .. code-block:: Python
+
+        # Remove all atoms with z coordinate >= 2.0
+        structure_without_Al = filter_structure_by_atom_info(
+            input_structure,
+            lambda atom_info: atom_info.atom_coordinate_cart[2] < 2.0,
+        )
+
+    Parameters
+    ----------
+    structure: _xtal.Structure
+        The initial structure
+    filter: Callable[[StructureAtomInfo], Any]
+        A function of `StructureAtomInfo` which returns True for atoms that should be
+        kept and False for atoms that should be removed.
+
+    Returns
+    -------
+    structure_after_removals: _xtal.Structure
+        A copy of `structure` excluding atoms for which the `filter` function returns
+        False.
+    """
+    return make_structure_from_atom_info(
+        lattice=structure.lattice(),
+        atoms=[x for x in make_structure_atom_info(structure) if filter(x)],
+        global_properties=structure.global_properties(),
+    )
+
+
+def combine_structures(
+    structures: list[_xtal.Structure],
+    lattice: Optional[_xtal.Lattice] = None,
+    global_properties: dict[str, np.ndarray[np.float64]] = {},
+) -> _xtal.Structure:
+    """Return a new structure which combines the species of all input structures
+
+    Parameters
+    ----------
+    structures: list[_xtal.Structure]
+        The structures to be combined.
+    lattice: Optional[_xtal.Lattice] = None
+        If not None, the lattice of the resulting structure. If None, the resulting
+        structure will use the lattice of the first structure.
+    global_properties: dict[str, numpy.ndarray[numpy.float64[m, n]]] = {}
+        Continuous properties associated with entire crystal, if present. Keys must be
+        the name of a CASM-supported property type. Values are (m, 1) arrays with
+        dimensions matching the standard dimension of the property type.
+
+    Returns
+    -------
+    combined_structure: _xtal.Structure
+        The resulting structure, which contains the species of all input structures
+        with Cartesian coordinates fixed to the same values as in the input structures.
+        All atom or molecule properties remain the same as in the input structures.
+    """
+    atoms = []
+    for structure in structures:
+        if lattice is None:
+            lattice = structure.lattice()
+        atoms += make_structure_atom_info(structure)
+    return make_structure_from_atom_info(
+        lattice=lattice,
+        atoms=atoms,
+        global_properties=global_properties,
     )